bar-archive-service/requirements.txt
backports.zoneinfo==0.2.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-tools==1.63.0
importlib-metadata==7.1.0
inflect==7.2.1
more-itertools==10.2.0
pika==1.3.2
protobuf==5.26.1
psycopg==3.1.19
psycopg2-binary==2.9.9
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
typeguard==4.2.1
typing-extensions==4.11.0
zipp==3.18.1
